gens_core/vdp/vdp_rend_m5.cpp: Ported GET_X_OFFSET to C++.
T_Get_X_Offset() is used to get the X offset for the current line using
the Horizontal Scroll Table and the current HScroll mode.

Like with T_Make_Sprite_Struct(), I added some C wrappers for the existing
assembly code to make porting everything easier.

+++ src/gens/gens_core/vdp/vdp_io.h
@@ -171,7 +171,7 @@ extern uint8_t *ScrA_Addr;
 extern uint8_t *ScrB_Addr;
 extern uint8_t *Win_Addr;
 extern uint8_t *Spr_Addr;
-extern uint8_t *H_Scroll_Addr;
+extern uint16_t *H_Scroll_Addr;
 
 // VDP convenience values: Horizontal.
 extern unsigned int H_Cell;
